[Step 5] 리뷰 반영 - 메서드, 클래스 위치
- winner 메서드를 TurnRecord로 이동
- dto 클래스 위치 이동
