+++ src/main/kotlin/racingcar/domain/carRacing/CarRecord.kt
@@ -1,0 +1,6 @@
+package racingcar.domain.carRacing
+
+data class CarRecord(
+    val car: Car,
+    val distance: Int
+)
