--- src/main/kotlin/racingcar/domain/carRacing/dto/CarRecord.kt
@@ -1,8 +1,0 @@
-package racingcar.domain.carRacing.dto
-
-import racingcar.domain.carRacing.Car
-
-data class CarRecord(
-    val car: Car,
-    val distance: Int
-)
