--- src/main/kotlin/racingcar/domain/carRacing/dto/TurnRecord.kt
@@ -1,6 +1,0 @@
-package racingcar.domain.carRacing.dto
-
-class TurnRecord(
-    val turn: Int,
-    val results: List<CarRecord>
-)
